Change frontend

+++ backend/src/service/article_service.py
@@ -1,9 +1,10 @@
 
-from turtle import up
 from typing import List
 
 from fastapi import HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.exc import NoResultFound
+
 from src.repository.article_repository import ArticleRepository
 from src.scheme.article import ArticleCreate, ArticleRead, ArticleUpdate
 
@@ -22,31 +23,39 @@ class ArticleService:
             return ArticleRead.model_validate(new_article, from_attributes=True)
         except Exception as e:
             await self.session.rollback()
-            raise HTTPException(status_code=400, detail=f"{e}")
+            raise HTTPException(status_code=400, detail=str(e))
         
     async def get_all(self, limit: int = 10, offset: int = 0) -> List[ArticleRead]:
-        return [ArticleRead.model_validate(i) for i in await self.repository.find_all(limit, offset)]
+        items = await self.repository.find_all(limit, offset)
+        return [ArticleRead.model_validate(i, from_attributes=True) for i in items]
 
     async def get_one_by_id(self, id: int) -> ArticleRead:
         find_article = await self.repository.find_one(id)
         if find_article:
-            return ArticleRead.model_validate(find_article)
+            return ArticleRead.model_validate(find_article, from_attributes=True)
         raise HTTPException(status_code=404, detail="Статья не найдена")
 
     async def delete_one(self, id: int):
         try:
+            obj = await self.repository.find_one(id)
+            if not obj:
+                raise
+            
             result = await self.repository.delete_article(id)
             await self.session.commit()
             return {"message": "Статья удалена"}
         except Exception as e:
             await self.session.rollback()
-            raise HTTPException(status_code=400, detail=f"Что-то случилось лол")
+            raise HTTPException(status_code=404, detail=str(e))
 
     async def update_one(self, id: int, article: ArticleUpdate) -> ArticleRead:
         try:
             updated_article: Article = await self.repository.update_one(id, article.model_dump(exclude_unset=True))
             await self.session.commit()
-            return ArticleRead.model_validate(updated_article)
+            return ArticleRead.model_validate(updated_article, from_attributes=True)
+        except NoResultFound:
+            await self.session.rollback()
+            raise HTTPException(status_code=404, detail=f"Статья с id={id} не найдена")
         except Exception as e:
             await self.session.rollback()
-            raise HTTPException(status_code=404, detail=f"Не существует тест по id={article.test_pk}")
+            raise HTTPException(status_code=400, detail=str(e))
